requirements.txt
altair==5.4.1
asn1crypto==1.5.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==43.0.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
platformdirs==4.3.6
protobuf==5.28.1
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
snowflake-connector-python
snowflake-snowpark-python
sortedcontainers==2.4.0
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
wheel==0.44.0
seaborn
geopandas
pyshp
openpyxl
vega_datasets
plotly
matplotlib
seaborn
requests
streamlit-mermaid